fix(kerykeion): extract PendingMessage/InflightMessage/LinkQuality behavior
TOPOLOGY/shallow-struct: PendingMessage and InflightMessage gain
is_expired(now) (and InflightMessage a has_timed_out(now) for the ACK
timeout), deduplicating the `now.duration_since(x.created) >= x.ttl`
check that was repeated at four call sites in OutboundQueue. LinkQuality
gains is_stale(cutoff), matching the fresh/stale cutoff comparison
already duplicated in remove_stale_nodes/remove_stale_links.

RUST/file-too-long: topology.rs's #[cfg(test)] block (338 lines) moves to
topology_tests.rs via the #[path = "..."] sibling-file convention already
used by collector_tests.rs/processor_tests.rs — 834 lines down to 496.

Refs #261
